refactor(home): drop Minimal layout; empty chat is the center prompt
Review feedback: Open to Home is the launch toggle — Home stays the full stats/recents panel. Remove homeLayout / homeSelectLatestFolder and HomeScreenMinimal. When Chat has no messages, show a centered composer with project picker (No project → home) under the pill; sidebar chrome remains. Settings copy updated accordingly.
