+++ src/renderer/src/components/settings-panel.tsx
@@ -75,12 +75,6 @@ export function SettingsPanel(): React.JSX.Element {
   const [autoScroll, setAutoScroll] = useState(draft0.autoScroll ?? settings?.autoScroll ?? DEFAULT_SETTINGS.autoScroll)
   const [resumeLastSession, setResumeLastSession] = useState(draft0.resumeLastSession ?? settings?.resumeLastSession ?? DEFAULT_SETTINGS.resumeLastSession)
   const [openToHomeOnLaunch, setOpenToHomeOnLaunch] = useState(draft0.openToHomeOnLaunch ?? settings?.openToHomeOnLaunch ?? DEFAULT_SETTINGS.openToHomeOnLaunch)
-  const [homeLayout, setHomeLayout] = useState<'info' | 'minimal'>(
-    draft0.homeLayout ?? settings?.homeLayout ?? DEFAULT_SETTINGS.homeLayout
-  )
-  const [homeSelectLatestFolder, setHomeSelectLatestFolder] = useState(
-    draft0.homeSelectLatestFolder ?? settings?.homeSelectLatestFolder ?? DEFAULT_SETTINGS.homeSelectLatestFolder
-  )
   const [runOnStartup, setRunOnStartup] = useState(draft0.runOnStartup ?? settings?.runOnStartup ?? DEFAULT_SETTINGS.runOnStartup)
   const [minimizeToTrayOnClose, setMinimizeToTrayOnClose] = useState(draft0.minimizeToTrayOnClose ?? settings?.minimizeToTrayOnClose ?? DEFAULT_SETTINGS.minimizeToTrayOnClose)
   const [permissionMode, setPermissionMode] = useState<PermissionMode>(
@@ -226,10 +220,6 @@ export function SettingsPanel(): React.JSX.Element {
     setAutoScroll(draft.autoScroll ?? settings.autoScroll)
     setResumeLastSession(draft.resumeLastSession ?? settings.resumeLastSession)
     setOpenToHomeOnLaunch(draft.openToHomeOnLaunch ?? settings.openToHomeOnLaunch)
-    setHomeLayout(draft.homeLayout ?? settings.homeLayout ?? DEFAULT_SETTINGS.homeLayout)
-    setHomeSelectLatestFolder(
-      draft.homeSelectLatestFolder ?? settings.homeSelectLatestFolder ?? DEFAULT_SETTINGS.homeSelectLatestFolder
-    )
     setRunOnStartup(draft.runOnStartup ?? settings.runOnStartup)
     setMinimizeToTrayOnClose(draft.minimizeToTrayOnClose ?? settings.minimizeToTrayOnClose)
     setPermissionMode(draft.permissionMode ?? settings.permissionMode)
@@ -437,8 +427,6 @@ export function SettingsPanel(): React.JSX.Element {
       autoScroll,
       resumeLastSession,
       openToHomeOnLaunch,
-      homeLayout,
-      homeSelectLatestFolder,
       runOnStartup,
       minimizeToTrayOnClose,
       permissionMode,
@@ -495,8 +483,6 @@ export function SettingsPanel(): React.JSX.Element {
       autoScroll: DEFAULT_SETTINGS.autoScroll,
       resumeLastSession: DEFAULT_SETTINGS.resumeLastSession,
       openToHomeOnLaunch: DEFAULT_SETTINGS.openToHomeOnLaunch,
-      homeLayout: DEFAULT_SETTINGS.homeLayout,
-      homeSelectLatestFolder: DEFAULT_SETTINGS.homeSelectLatestFolder,
       runOnStartup: DEFAULT_SETTINGS.runOnStartup,
       minimizeToTrayOnClose: DEFAULT_SETTINGS.minimizeToTrayOnClose,
       permissionMode: DEFAULT_SETTINGS.permissionMode,
@@ -511,8 +497,6 @@ export function SettingsPanel(): React.JSX.Element {
     setAutoScroll(defaults.autoScroll!)
     setResumeLastSession(defaults.resumeLastSession!)
     setOpenToHomeOnLaunch(defaults.openToHomeOnLaunch!)
-    setHomeLayout(defaults.homeLayout!)
-    setHomeSelectLatestFolder(defaults.homeSelectLatestFolder!)
     setRunOnStartup(defaults.runOnStartup!)
     setMinimizeToTrayOnClose(defaults.minimizeToTrayOnClose!)
     setPermissionMode(defaults.permissionMode!)
@@ -781,54 +765,10 @@ export function SettingsPanel(): React.JSX.Element {
 
           <SettingsRow
             label="Open to Home Screen on Launch"
-            description="Show the Home launcher on startup; Pi starts only when you open a workspace or session"
+            description="On: full Home launcher (stats, recents, open folder). Off: open Chat with the empty-session center prompt and project picker."
           >
             <Toggle checked={openToHomeOnLaunch} onChange={(v) => { setOpenToHomeOnLaunch(v); setSettingsDraft({ openToHomeOnLaunch: v }) }} />
           </SettingsRow>
-
-          <SettingsRow
-            label="Home Layout"
-            description="Info: full Home panel (stats, recents, open folder). Minimal: center prompt with project picker (sidebar stays visible)."
-          >
-            <div className="ml-auto inline-flex w-fit gap-0.5 rounded-md bg-card/60 p-0.5">
-              {([
-                { id: 'info' as const, label: 'Info' },
-                { id: 'minimal' as const, label: 'Minimal' },
-              ]).map((opt) => (
-                <button
-                  key={opt.id}
-                  type="button"
-                  onClick={() => {
-                    setHomeLayout(opt.id)
-                    setSettingsDraft({ homeLayout: opt.id })
-                    void applyImmediate({ homeLayout: opt.id })
-                  }}
-                  className={clsx(
-                    'rounded px-2.5 py-1 text-xs font-medium transition-colors',
-                    homeLayout === opt.id ? 'bg-elevated text-primary' : 'text-dim hover:text-secondary'
-                  )}
-                >
-                  {opt.label}
-                </button>
-              ))}
-            </div>
-          </SettingsRow>
-
-          {homeLayout === 'minimal' && (
-            <SettingsRow
-              label="Select latest folder"
-              description="On Minimal Home, pre-select your most recently used project. Off = start on No project (home directory)."
-            >
-              <Toggle
-                checked={homeSelectLatestFolder}
-                onChange={(v) => {
-                  setHomeSelectLatestFolder(v)
-                  setSettingsDraft({ homeSelectLatestFolder: v })
-                  void applyImmediate({ homeSelectLatestFolder: v })
-                }}
-              />
-            </SettingsRow>
-          )}
 
           <SettingsRow
             label="Resume Last Session"
